Fix cells (add UserTableViewCell) and fix icons (add view).
