+++ SettingsScreenApplication/Sources/MainModule/ViewController.swift
@@ -4,7 +4,6 @@
 //
 //  Created by Nika Semenkova on 01.02.2023.
 //
-
 import UIKit
 import SnapKit
 
@@ -25,7 +24,8 @@ class ViewController: UIViewController {
     
     private lazy var tableView: UITableView = {
         let tableView = UITableView(frame: .zero, style: .insetGrouped)
-        tableView.register(TableViewCell.self, forCellReuseIdentifier: TableViewCell.identifier)
+        tableView.register(DefaultTableViewCell.self, forCellReuseIdentifier: DefaultTableViewCell.identifier)
+        tableView.register(UserTableViewCell.self, forCellReuseIdentifier: UserTableViewCell.identifier)
         tableView.dataSource = self
         tableView.delegate = self
         tableView.translatesAutoresizingMaskIntoConstraints = false
@@ -63,7 +63,6 @@ class ViewController: UIViewController {
 }
 
 // MARK: - Extensions
-
 extension ViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         if model?[indexPath.section] == model?[0] { return 80 } else { return 40 }
@@ -78,22 +77,41 @@ extension ViewController: UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: TableViewCell.identifier, for: indexPath) as? TableViewCell
-        cell?.cells = model?[indexPath.section][indexPath.row]
-        
-        switch cell?.cells?.icon {
-        case "airplane":
-            let switchView = UISwitch(frame: .zero)
-            switchView.setOn(false, animated: true)
-            cell?.accessoryView = switchView
+        switch model?[indexPath.section] {
+        case model?[0]:
+            let cell = tableView.dequeueReusableCell(withIdentifier: UserTableViewCell.identifier, for: indexPath) as? UserTableViewCell
+            cell?.cells = model?[indexPath.section][indexPath.row]
+            return cell ?? UITableViewCell()
         default:
-            cell?.accessoryType = .disclosureIndicator
+            let cell = tableView.dequeueReusableCell(withIdentifier: DefaultTableViewCell.identifier, for: indexPath) as? DefaultTableViewCell
+            cell?.cells = model?[indexPath.section][indexPath.row]
+            
+            switch cell?.cells?.icon {
+            case "airplane":
+                let switchView = UISwitch(frame: .zero)
+                switchView.setOn(false, animated: true)
+                cell?.accessoryView = switchView
+            default:
+                cell?.accessoryType = .disclosureIndicator
+            }
+            
+            return cell ?? UITableViewCell()
         }
+        //        let cell = tableView.dequeueReusableCell(withIdentifier: DefaultTableViewCell.identifier, for: indexPath) as? DefaultTableViewCell
+        //        cell?.cells = model?[indexPath.section][indexPath.row]
+        //
+        //        switch cell?.cells?.icon {
+        //        case "airplane":
+        //            let switchView = UISwitch(frame: .zero)
+        //            switchView.setOn(false, animated: true)
+        //            cell?.accessoryView = switchView
+        //        default:
+        //            cell?.accessoryType = .disclosureIndicator
+        //        }
         
-        return cell ?? UITableViewCell()
+        //        return cell ?? UITableViewCell()
     }
 }
-
 extension ViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
@@ -102,4 +120,3 @@ extension ViewController: UITableViewDelegate {
         navigationController?.pushViewController(detailView, animated: true)
     }
 }
-
